refactor: change path to password-recovery widgets
