+++ src/app/auth/forgot-password/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { ForgotPassword } from '@/widgets/(password-recovery)/forgot-password-section'
+import { ForgotPassword } from '@/widgets/password-recovery/forgot-password-section'
 
 const page = () => {
   return <ForgotPassword />
