+++ src/app/auth/password-recovery/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { CreatePassword } from '@/widgets/(password-recovery)/create-new-password-section'
+import { CreatePassword } from '@/widgets/password-recovery/create-new-password-section'
 
 const page = () => {
   return <CreatePassword />
