+++ src/app/auth/verification-link-expired/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { LinkExpired } from '@/widgets/(password-recovery)/link-expired-section'
+import { LinkExpired } from '@/widgets/password-recovery/link-expired-section'
 
 const page = () => {
   return <LinkExpired />
