+++ src/widgets/password-recovery/create-new-password-section/CreatePassword.module.scss
