+++ src/widgets/password-recovery/forgot-password-section/ForgotPassword.module.scss
