+++ src/widgets/password-recovery/link-expired-section/LinkExpired.module.scss
